+++ CHANGELOG.rst
@@ -6,10 +6,18 @@ dcicutils
 Change Log
 ----------
 
-6.9.0
+6.10.0
 =====
 
 * Move ``mocked_s3utils_with_sse`` from ``test_ff_utils.py`` to ``ff_mocks.py``.
+
+
+6.9.0
+=====
+
+* In ``dcicutils.misc_utils``:
+
+  * Add method ``is_c4_arn`` to check if given ARN looks like CGAP or Fourfront entity.
 
 
 6.8.0
@@ -18,6 +26,7 @@ Change Log
 * In `dcicutils.deployment_utils``
 
   * Add support for ``Auth0Domain`` and ``Auth0AllowedConnections``
+
 
 
 6.7.0